Plot r0 by location as extracted from hierarchical outbreak model. If extracted_posts is NULL then posts object is used to first extract the R0 and incidence draws from posterior (note this will take longer)

Plot r0 by location as extracted from hierarchical outbreak model. If extracted_posts is NULL then posts object is used to first extract the R0 and incidence draws from posterior (note this will take longer)

hom_plot_r0_by_location(extracted_posts = NULL, posts = NULL)

Arguments

extracted_posts
object returned by hom_extract_posterior_draws

posts
Object after calling extract of stan model object of hierarchical model

Value

list containing: * plot - ggplot object * table - tibble object of results

Note

Naming convention throughout is snake case with prefix "hom_" to denote Hierarcical Outbreak Model

Examples

 mod # RSTAN fit object

#> Inference for Stan model: hierarchical_SEIR_incidence_model.
#> 4 chains, each with iter=600; warmup=300; thin=1; 
#> post-warmup draws per chain=300, total post-warmup draws=1200.
#> 
#>                     mean se_mean   sd   2.5%   25%    50%    75%  97.5% n_eff
#> y0[1]               0.90    0.00 0.01   0.88  0.89   0.90   0.90   0.92  1194
#> y0[2]               0.10    0.00 0.01   0.08  0.10   0.10   0.11   0.12  1194
#> y0[3]               0.00     NaN 0.00   0.00  0.00   0.00   0.00   0.00   NaN
#> y0[4]               0.00     NaN 0.00   0.00  0.00   0.00   0.00   0.00   NaN
#> y0[5]               0.00     NaN 0.00   0.00  0.00   0.00   0.00   0.00   NaN
#> params[1]           4.90    0.04 1.44   2.63  3.87   4.74   5.77   8.49  1104
#> params[2]           0.21    0.00 0.02   0.16  0.19   0.21   0.22   0.25  1374
#> params[3]           0.13    0.00 0.01   0.10  0.12   0.13   0.14   0.15  1062
#> params[4]         103.59    0.24 9.88  84.35 96.92 103.69 110.11 122.70  1630
#> params[5]           0.92    0.02 0.66   0.04  0.39   0.81   1.34   2.43   869
#> params[6]           0.09    0.00 0.07   0.00  0.04   0.08   0.13   0.26  1267
#> r0k[1]              4.61    0.04 1.33   2.45  3.67   4.49   5.31   7.59  1244
#> r0k[2]              4.90    0.04 1.44   2.63  3.87   4.74   5.77   8.49  1104
#> r0                  3.69    0.03 0.92   1.92  3.06   3.64   4.35   5.56  1248
#> r0_sigma            1.07    0.02 0.69   0.09  0.51   0.97   1.52   2.62  1033
#> incidence[1,1]      1.91    0.01 0.29   1.37  1.72   1.89   2.09   2.53  1485
#> incidence[1,2]      2.00    0.01 0.32   1.42  1.79   1.99   2.19   2.66  1647
#> incidence[2,1]      3.63    0.01 0.53   2.64  3.26   3.59   3.95   4.74  1513
#> incidence[2,2]      3.82    0.01 0.59   2.69  3.42   3.79   4.16   5.06  1711
#> incidence[3,1]      5.30    0.02 0.77   3.88  4.77   5.27   5.77   6.96  1558
#> incidence[3,2]      5.60    0.02 0.87   3.94  5.00   5.57   6.12   7.46  1803
#> incidence[4,1]      7.01    0.03 1.05   5.14  6.29   6.95   7.66   9.20  1608
#> incidence[4,2]      7.44    0.03 1.20   5.21  6.65   7.37   8.16  10.02  1889
#> incidence[5,1]      8.79    0.03 1.39   6.25  7.83   8.70   9.65  11.79  1646
#> incidence[5,2]      9.38    0.04 1.61   6.44  8.34   9.27  10.32  12.86  1935
#> hyper_priors[1]     2.96    0.03 0.96   1.07  2.30   2.91   3.65   4.86  1339
#> hyper_priors[2]     0.20    0.00 0.01   0.18  0.19   0.20   0.21   0.22  1272
#> hyper_priors[3]     0.17    0.00 0.01   0.16  0.16   0.17   0.17   0.18  1110
#> hyper_priors[4]     0.81    0.02 0.60   0.04  0.31   0.69   1.18   2.21  1135
#> hyper_priors[5]     0.13    0.00 0.08   0.01  0.07   0.12   0.18   0.29  1055
#> fake_incidence[1]   0.00     NaN 0.00   0.00  0.00   0.00   0.00   0.00   NaN
#> fake_incidence[2]   1.86    0.04 1.44   0.00  1.00   2.00   3.00   5.00  1044
#> fake_incidence[3]   1.78    0.04 1.36   0.00  1.00   2.00   3.00   5.00  1236
#> fake_incidence[4]   1.87    0.04 1.43   0.00  1.00   2.00   3.00   5.00  1125
#> fake_incidence[5]   1.98    0.04 1.46   0.00  1.00   2.00   3.00   5.02  1328
#> predictive_r0       3.72    0.05 1.47   0.78  2.80   3.71   4.64   6.79  1017
#> lp__               -7.68    0.15 3.01 -14.45 -9.51  -7.23  -5.57  -2.76   405
#>                   Rhat
#> y0[1]                1
#> y0[2]                1
#> y0[3]              NaN
#> y0[4]              NaN
#> y0[5]              NaN
#> params[1]            1
#> params[2]            1
#> params[3]            1
#> params[4]            1
#> params[5]            1
#> params[6]            1
#> r0k[1]               1
#> r0k[2]               1
#> r0                   1
#> r0_sigma             1
#> incidence[1,1]       1
#> incidence[1,2]       1
#> incidence[2,1]       1
#> incidence[2,2]       1
#> incidence[3,1]       1
#> incidence[3,2]       1
#> incidence[4,1]       1
#> incidence[4,2]       1
#> incidence[5,1]       1
#> incidence[5,2]       1
#> hyper_priors[1]      1
#> hyper_priors[2]      1
#> hyper_priors[3]      1
#> hyper_priors[4]      1
#> hyper_priors[5]      1
#> fake_incidence[1]  NaN
#> fake_incidence[2]    1
#> fake_incidence[3]    1
#> fake_incidence[4]    1
#> fake_incidence[5]    1
#> predictive_r0        1
#> lp__                 1
#> 
#> Samples were drawn using NUTS(diag_e) at Thu Jan 07 19:22:05 2021.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).

 posts <- rstan::extract(mod) # extract posterior from model object
 extracted_posts <- hom_extract_posterior_draws(posts) # get object of incidence and r0

#> Error in tibble::as_tibble(posts$r0k) %>% tidyr::pivot_longer(dplyr::everything(),     names_to = "location", values_to = "r0") %>% dplyr::mutate(location = str_remove_all(location,     "V")): could not find function "%>%"

 result <- hom_plot_r0_by_location(extracted_posts=extracted_posts)

#> Error in hom_plot_r0_by_location(extracted_posts = extracted_posts): object 'extracted_posts' not found

 # plot results
 show(result$plot)

#> Error in show(result$plot): object 'result' not found